+++ CHANGES.md
@@ -1,2 +1,5 @@
 Changes
 ===================
+###2015.09.21
+* Added default context
+* Added command-specific contexts